fixed board constructor and unit test object creation

+++ CSCI-4526-Sudoku/P5-KorideMok/Board-KorideMok.cpp
@@ -9,7 +9,7 @@ static const char* clusterT[3] = {"ROW", "COLUMN", "BOX"};
 // Preconditions: Game object exists
 // Postconditions: Board object is created
 Board::
-Board(int n, ifstream& puzfile) : n(n), file(puzfile), left(n*n) {
+Board(short n, ifstream& puzfile) : n(n), file(puzfile), left(n*n) {
     bd = new Square[n*n];
     getPuzzle();
     makeClusters();
